+++ src/main/java/com/reviewer/entity/Review.java
@@ -11,6 +11,7 @@ import org.springframework.data.mongodb.core.index.CompoundIndex;
 import org.springframework.data.mongodb.core.index.CompoundIndexes;
 import org.springframework.data.mongodb.core.index.Indexed;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 import java.util.UUID;
 
@@ -40,5 +41,5 @@ public class Review {
 
     private Boolean isActive = false;
 
-    private Float average;
+    private BigDecimal average;
 }
